:hammer: Refactor 과제 삭제 시 과제 작성자 id와 userID 비교후 삭제

+++ src/main/java/Umc/replendar/assignment/service/AssignmentServiceImpl.java
@@ -260,10 +260,15 @@ public class AssignmentServiceImpl implements AssignmentService {
 
     //활동로그에서 해당 과제가 다 삭제되는지 확인해야함 - 삭제됨
     @Override
-    public ApiResponse<String> deleteAssignment(Long assId) {
+    public ApiResponse<String> deleteAssignment(Long assId, Long userId) {
+
+        User user = userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
 
         Assignment assignment = assignmentRepository.findById(assId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 과제입니다."));
 
+        if(!user.getId().equals(assignment.getUser().getId())){
+            return ApiResponse.onFailure("INVALID_REQUEST", "본인의 과제만 삭제할 수 있습니다.", null);
+        }
         assignmentRepository.delete(assignment);
 
         return ApiResponse.onSuccess("과제가 삭제되었습니다.");
